refactor: extract task view helpers and isolate momentum logic
