+++ src/constants/TaskRanks.jsx
@@ -1,0 +1,11 @@
+export const LOAD_RANK = {
+  low: 0,
+  medium: 1,
+  high: 2,
+};
+
+export const PRIORITY_RANK = {
+  low: 0,
+  medium: 1,
+  high: 2,
+};
